moved UnmarshalTraces to pkg/translator/azure
